add referral system and subscribtion status

+++ src/internals/app/admin/command/removeAdmin.ts
@@ -1,3 +1,4 @@
+import { BadRequestError } from "../../../../pkg/errors/customError";
 import { AdminRepository } from "../../../domain/admins/repository";
 
 export interface RemoveAdminCommand {
@@ -15,6 +16,10 @@ export class RemoveAdminCommandC implements RemoveAdminCommand {
 
     Handle = async (adminId: string): Promise<void> => {
         try {
+            const admin  = await this.adminRepository.GetAdminByID(adminId)
+            if (admin?.examAccess?.includes("SUPER_ADMIN")) {
+                throw  new BadRequestError("Super Admin can not be deleted")
+            }
             await this.adminRepository.RemoveAdmin(adminId)
         } catch (error) {
             throw error;
